simplifying places, splitting into more files
