+++ bigraph_schema/bigraph.py
@@ -1,4 +1,6 @@
 from pprint import pformat as pf
+from bigraph_schema.bigraph.basis import Joindict
+from bigraph_schema.bigraph.placement import PlacementPlaces
 
 """
 attempting a normalized version of the process bigraph abstraction
@@ -34,57 +36,6 @@ import uuid
 
 def new_uuid():
     return uuid.uuid4()
-
-
-class JoinDict():
-    """
-    many to many mapping between two cachable types
-    """
-    def __init__(self):
-        self.left = {}
-        self.right = {}
-
-    def _insert_pair(self, place, a, b):
-        if not place.get(a):
-            place[a] = {}
-        place[a][b] = True
-
-    def insert(self, left, right):
-        self._insert_pair(self.left, left, right)
-        self._insert_pair(self.right, right, left)
-
-    def _purge_token(self, a, b, token):
-        for token_b in a[token].keys():
-            b[token_b].pop(token, None)
-        a.pop(token, None)
-
-    def remove_left(self, left):
-        self._purge_token(self.left, self.right, left)
-
-    def remove_right(self, right):
-        self._purge_token(self.right, self.left, right)
-
-    def delink(self, left, right):
-        try:
-            self.right[right].pop(left)
-        except KeyError:
-            pass
-        try:
-            self.left[left].pop(right)
-        except KeyError:
-            pass
-
-    def get_left(self, item):
-        result = self.left.get(item) or {}
-        return list(result.keys())
-
-    def get_right(self, item):
-        result = self.right.get(item) or {}
-        return list(result.keys())
-
-    def update(self, other):
-        self.left.update(other.left)
-        self.right.update(other.right)
 
 
 class Link():
@@ -180,239 +131,6 @@ class Link():
             result.append(self.reaction(self, bigraph, reactum))
         return result
 
-
-class Place():
-    """
-    defines a place in terms of parent (outer) and list of child (inner) nodes
-    within a tree, a value of None in either position indicates an interface
-    """
-    def __init__(self, node_id, inner=None, outer=None):
-        self.node_id = node_id
-        # list of place_id or None
-        self.inner = inner or []
-        # single value of place_id or None
-        self.outer = outer
-
-    def __repr__(self):
-        return f'Place({node_id}, inner={self.inner}, outer={self.outer})'
-
-    def __eq__(self, other):
-        return self.inner == other.inner and self.outer == other.outer
-
-class Places():
-    def as_tree(self):
-        roots = self.get_roots()
-
-        result = {}
-        for root in roots:
-            result[root] = self.build_tree(root)
-
-        return result
-
-    def add_branch(self, parent, child):
-        pass
-    def get_branches(self, node_id):
-        pass
-    def get_place(self, place_id):
-        pass
-    def insert(self, place):
-        pass
-    def get_roots(self):
-        pass
-    def get_leaves(self):
-        pass
-    def build_tree(self, root):
-        """
-        returns a tree of node_id down from root
-        """
-        pass
-
-class PlacePlaces(Places):
-    def __init__(self):
-        # index of node to place
-        self.db = {}
-
-    def insert(self, place):
-        self.db[place.node_id] = place
-
-    def add_branch(self, parent, child):
-        existing_child = self.db.get(child)
-        if not existing_child:
-            self.insert(Place(child, outer=parent))
-        existing_parent = self.db.get(parent)
-        if not existing_parent:
-            existing_parent = Place(parent, inner=[child])
-            self.insert(existing_parent)
-        if not child in existing_parent.inner:
-            existing_parent.inner.append(child)
-
-    def get_branches(self, node_id):
-        return self.db[node_id].inner
-
-    def get_place(self, node_id):
-        place = self.db.get(node_id)
-        return {'node': place.node_id,
-                'inner': place.inner,
-                'outer': place.outer}
-
-    def get_roots(self):
-        roots = []
-        for node_id,place in self.db.items():
-            if place.outer == None:
-                roots.append(node_id)
-        return roots
-
-    def get_leaves(self):
-        result = {}
-        for uid,place in self.db.items():
-            if place.inner == None:
-                result[uid] = place
-        return result
-
-    def build_tree(self, root):
-        children = self.db[root].inner or []
-        result = {}
-        for child_id in children:
-            child = self.db[child_id]
-            if child.inner == None:
-                result[child_id] = None
-            else:
-                result[child_id] = self.build_tree(child_id)
-        return result
-
-
-class PlacementPlaces(Places):
-    def __init__(self):
-        self.placements = {}
-        self.placements_nodes = JoinDict()
-
-    def __repr__(self):
-        return f'PlacementPlaces(placements={self.placements})'
-
-    def insert(self, placement):
-        self.placements[placement.id] = placement
-        self.placements_nodes.insert(placement.id, placement.node_id)
-
-    def add_branch(self, parent_node, child_node):
-        """
-        parent_node, child_node: have type node_id
-        """
-        child_placements = self.placements_nodes.get_right(child_node)
-        for place_id in child_placements:
-            if self.placements[place_id].outer == parent_node:
-                return
-        self.insert(Placement(child_node, outer=parent_node))
-
-    def get_place(self, node_id):
-        placement_ids = self.placements_nodes.get_right(node_id)
-        result = {'node_id': node_id, 'inner': []}
-        for placement_id in placement_ids:
-            placement = self.placements.get(placement_id)
-            if placement.outer and placement.outer != node_id:
-                if result.get('outer'):
-                    raise RuntimeError("multiple parent nodes in tree", {
-                                           'node_id': node_id,
-                                           'a': result['outer'],
-                                           'b': placement.outer})
-                result['outer'] = placement.outer
-            if placement.outer == node_id:
-                result['inner'].append(placment.inner)
-        result['outer'] = result.get('outer') or None
-        return result
-
-    def get_branches(self, node_id):
-        # returns a list of uuids
-        placements = self.placements_nodes.get_right(node_id)
-        result = [self.placements[placement] for placement in placements]
-        return result
-
-    def __get_node(self, placement_id):
-        nodes = self.placements_nodes.get_left(placement_id)
-        assert len(nodes) <= 1, \
-                f"placements must have 0 or 1 node, got: {nodes}"
-        if nodes == []:
-            return None
-        else:
-            return nodes[0]
-
-    def get_roots(self):
-        roots = []
-        for placement_id,placement in self.placements.items():
-            if placement.outer == None:
-                roots.append(self.__get_node(placement_id))
-        return roots
-
-    def get_leaves(self):
-        result = {}
-        for placement_id,placement in self.placements():
-            if placement.inner == None:
-                result[self.__get_node(placment_id)] = placement
-        return result
-
-    def build_tree(self, root):
-        branches = self.placements_nodes.get_right(root)
-        result = {}
-        for branch in branches:
-            if branch.inner == None:
-                result[branch.node_id] = None
-            else:
-                resut[branch.node_id] = self.build_tree(branch.node_id)
-        return result
-
-class Placement():
-    """
-    defines a placement in terms of a parent and a child place. multiple place
-    objects will represent the same parent node if it has multiple children
-
-    a value of None in either position indicates an interface
-    """
-    def __init__(self, node_id, uid=None, inner=None, outer=None):
-        self.id = uid or new_uuid()
-        self.node_id = node_id
-        self.outer = outer
-
-    def __repr__(self):
-        return f'Placement(node_id={self.node_id}, ' \
-               f'inner={self.inner}, ' \
-               f'outer={self.outer})'
-
-    def __eq__(self, other):
-        return [self.node_id, self.inner, self.outer] == \
-                [other.node_id, other.inner, other.outer]
-
-#    def get_roots(places):
-#        """
-#        takes a mapping from node_id to list of placements
-#
-#        returns a list of placements that are roots
-#        """
-#        roots = []
-#        for node_id, placements in places.items():
-#            for placement in placements:
-#                if placement.outer == None:
-#                    roots.append(placement)
-#        return roots
-#
-#    def build_tree(places, root):
-#        """
-#        takes a mapping from node_id to list of placements
-#
-#        returns a tree of node_id
-#
-#        errors if any node_id has more than one parent
-#        """
-#        child_places = places[root]
-#        children = [placement.node_id for placement in child_places]
-#        result = {}
-#        for child_id in children:
-#            child_placements = places[child_id]
-#            if child_placements == []:
-#                result[child_id] == None
-#            else:
-#                for placement in child_placements:
-#                    node = placement.node_id
-#                    result[node] = build_tree(places, node)
-#        return result
 
 
 class Bigraph():
